@@ -1,9 +1,7 @@
 ﻿using ErrorOr;
 using KittyStore.Domain.ShopCartAggregate;
-using KittyStore.Domain.UserAggregate.ValueObjects;
 using MediatR;
 
-namespace KittyStore.Application.ShopCarts.Queries.GetShopCart
-{
-    public record GetShopCartQuery(UserId UserId) : IRequest<ErrorOr<ShopCart>>;
-}
+namespace KittyStore.Application.ShopCarts.Queries.GetShopCart;
+
+public record GetShopCartQuery() : IRequest<ErrorOr<ShopCart>>;
