@@ -7,8 +7,6 @@ using KittyStore.Application.Users.Queries.GetAllUsers;
 using KittyStore.Application.Users.Queries.GetUserById;
 using KittyStore.Contracts.Admin.Cats;
 using KittyStore.Contracts.Admin.Users;
-using KittyStore.Domain.CatAggregate.ValueObjects;
-using KittyStore.Domain.UserAggregate.ValueObjects;
 using MapsterMapper;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
@@ -25,31 +23,30 @@ namespace KittyStore.Api.Controllers
         [HttpPost("cats")]
         public async Task<IActionResult> CreateCat(CreateCatRequest request)
         {
-            var command = _mapper.Map<CreateCatCommand>(request);
-            var createdResult = await _mediator.Send(command);
+            var command = Mapper.Map<CreateCatCommand>(request);
+            var createdResult = await Mediator.Send(command);
         
             return createdResult.Match(
-                cat => Ok(_mapper.Map<CatResponse>(cat)),
+                cat => Ok(Mapper.Map<CatResponse>(cat)),
                 errors => Problem(errors));
         }
     
         [HttpPut("cats/{id:guid}")]
         public async Task<IActionResult> UpdateCat(Guid id, UpdateCatRequest request)
         {
-            var command = _mapper.Map<UpdateCatCommand>(request);
-            command.Id = CatId.Create(id);
-            var updatedResult = await _mediator.Send(command);
+            var command = Mapper.Map<UpdateCatCommand>(request);
+            command.Id = id;
+            var updatedResult = await Mediator.Send(command);
         
             return updatedResult.Match(
-                cat => Ok(_mapper.Map<CatResponse>(cat)),
+                cat => Ok(Mapper.Map<CatResponse>(cat)),
                 errors => Problem(errors));
         }
     
         [HttpDelete("cats/{id:guid}")]
         public async Task<IActionResult> DeleteCat(Guid id)
         {
-            var catId = CatId.Create(id);
-            var deletedResult = await _mediator.Send(new DeleteCatCommand(catId));
+            var deletedResult = await Mediator.Send(new DeleteCatCommand(id));
         
             return deletedResult.Match(
                 _ => Ok(),
@@ -59,41 +56,40 @@ namespace KittyStore.Api.Controllers
         [HttpGet("users")]
         public async Task<IActionResult> GetAllUsers()
         {
-            var result = await _mediator.Send(new GetAllUsersQuery());
+            var result = await Mediator.Send(new GetAllUsersQuery());
 
             return result.Match(
-                users => Ok(_mapper.Map<List<UserResponse>>(users)),
+                users => Ok(Mapper.Map<List<UserResponse>>(users)),
                 errors => Problem(errors));
         }
         
         [HttpGet("users/{id:guid}")]
         public async Task<IActionResult> GetUser(Guid id)
         {
-            var result = await _mediator.Send(
-                new GetUserByIdQuery(UserId.Create(id)));
+            var result = await Mediator.Send(
+                new GetUserByIdQuery(id));
 
             return result.Match(
-                user => Ok(_mapper.Map<UserResponse>(user)),
+                user => Ok(Mapper.Map<UserResponse>(user)),
                 errors => Problem(errors));
         }
     
         [HttpPut("users/{id:guid}")]
         public async Task<IActionResult> UpdateUser(Guid id, UpdateUserRequest request)
         {
-            var command = _mapper.Map<UpdateUserCommand>(request);
-            command.Id = UserId.Create(id);
-            var updatedResult = await _mediator.Send(command);
+            var command = Mapper.Map<UpdateUserCommand>(request);
+            command.Id = id;
+            var updatedResult = await Mediator.Send(command);
 
             return updatedResult.Match(
-                user => Ok(_mapper.Map<UserResponse>(user)),
+                user => Ok(Mapper.Map<UserResponse>(user)),
                 errors => Problem(errors));
         }
     
         [HttpDelete("users/{id:guid}")]
         public async Task<IActionResult> DeleteUser(Guid id)
         {
-            var userId = UserId.Create(id);
-            var deletedResult = await _mediator.Send(new DeleteUserCommand(userId));
+            var deletedResult = await Mediator.Send(new DeleteUserCommand(id));
         
             return deletedResult.Match(
                 _ => Ok(),
