@@ -15,24 +15,24 @@ namespace KittyStore.Api.Controllers
         public AuthenticationController(ISender mediator, IMapper mapper): base(mediator, mapper) { }
 
         [HttpPost("register")]
-        public async Task<IActionResult> Register([FromForm]RegisterRequest request)
+        public async Task<IActionResult> Register(RegisterRequest request)
         {
-            var command = _mapper.Map<RegisterCommand>(request);
-            var registerResult = await _mediator.Send(command);
+            var command = Mapper.Map<RegisterCommand>(request);
+            var registerResult = await Mediator.Send(command);
 
             return registerResult.Match(
-                authResult => Ok(_mapper.Map<AuthenticationResponse>(authResult)),
+                authResult => Ok(Mapper.Map<AuthenticationResponse>(authResult)),
                 errors => Problem(errors));
         }
 
         [HttpPost("login")]
-        public async Task<IActionResult> Login([FromForm]LoginRequest request)
+        public async Task<IActionResult> Login(LoginRequest request)
         {
-            var command = _mapper.Map<LoginQuery>(request);
-            var loginResult = await _mediator.Send(command);
+            var command = Mapper.Map<LoginQuery>(request);
+            var loginResult = await Mediator.Send(command);
 
             return loginResult.Match(
-                authResult => Ok(_mapper.Map<AuthenticationResponse>(authResult)),
+                authResult => Ok(Mapper.Map<AuthenticationResponse>(authResult)),
                 errors => Problem(errors));
         }
     }
