@@ -22,7 +22,7 @@ var app = builder.Build();
     
     using var scope = app.Services.CreateScope();
     var db = scope.ServiceProvider.GetRequiredService<AppDbContext>();
-    db.Database.EnsureDeleted();
+    //db.Database.EnsureDeleted();
     db.Database.EnsureCreated();
     
     app.UseExceptionHandler("/error");
