@@ -1,10 +1,9 @@
-﻿using System.Security.Cryptography;
-using System.Text;
-using MediatR;
+﻿using MediatR;
 using ErrorOr;
 using KittyStore.Application.Authentication.Common;
 using KittyStore.Application.Common.Interfaces.Authentication;
 using KittyStore.Application.Common.Interfaces.Persistence;
+using KittyStore.Application.Common.Interfaces.Utils;
 using KittyStore.Domain.Common.Errors;
 
 namespace KittyStore.Application.Authentication.Queries.Login
@@ -13,35 +12,32 @@ namespace KittyStore.Application.Authentication.Queries.Login
     {
         private readonly IUserRepository _userRepository;
         private readonly IJwtTokenGenerator _jwtTokenGenerator;
-    
-        public LoginQueryHandler(IUserRepository userRepository, IJwtTokenGenerator jwtTokenGenerator)
+        private readonly IPasswordService _passwordService;
+
+        public LoginQueryHandler(IUserRepository userRepository, IJwtTokenGenerator jwtTokenGenerator,
+            IPasswordService passwordService)
         {
             _userRepository = userRepository;
             _jwtTokenGenerator = jwtTokenGenerator;
+            _passwordService = passwordService;
         }
-    
+
         public async Task<ErrorOr<AuthenticationResult>> Handle(LoginQuery query, CancellationToken cancellationToken)
         {
             //Validate parameters
             if (await _userRepository.GetUserByEmailAsync(query.Email) is not { } user)
                 return Errors.Authentication.InvalidCredentials;
 
-            if (!VerifyPasswordHash(query.Password, user.PasswordHash, user.PasswordSalt))
+            var isSuccess = _passwordService.VerifyPassword(query.Password, user.PasswordHash, user.PasswordSalt);
+            if (!isSuccess)
                 return new[] { Errors.Authentication.InvalidCredentials };
 
             //Generate token
             var token = _jwtTokenGenerator.GenerateToken(user);
-        
-            return new AuthenticationResult(
-                user, 
-                token);
-        }
 
-        private bool VerifyPasswordHash(string password, byte[] passwordHash, byte[] passwordSalt)
-        {
-            using var hmac = new HMACSHA512(passwordSalt);
-            var computedHash = hmac.ComputeHash(Encoding.UTF8.GetBytes(password));
-            return computedHash.SequenceEqual(passwordHash);
+            return new AuthenticationResult(
+                user,
+                token);
         }
     }
 }
