@@ -1,8 +1,7 @@
 ﻿using ErrorOr;
-using KittyStore.Domain.CatAggregate.ValueObjects;
 using MediatR;
 
 namespace KittyStore.Application.Cats.Commands.DeleteCat
 {
-    public record DeleteCatCommand(CatId Id) : IRequest<ErrorOr<Unit>>;
+    public record DeleteCatCommand(Guid Id) : IRequest<ErrorOr<Unit>>;
 }
