@@ -2,7 +2,7 @@
 {
     public interface ICacheService
     {
-        Task<T> GetDataAsync<T>(string key);
+        Task<T?> GetDataAsync<T>(string key);
     
         Task SetDataAsync<T>(string key, T value, DateTimeOffset expirationTime);
     
