@@ -1,11 +1,10 @@
 ﻿using KittyStore.Domain.OrderAggregate;
-using KittyStore.Domain.UserAggregate.ValueObjects;
 
 namespace KittyStore.Application.Common.Interfaces.Persistence
 {
     public interface IOrderRepository
     {
-        Task<List<Order>> GetUserOrdersAsync(UserId id);
+        Task<List<Order>> GetUserOrdersAsync(Guid id);
 
         Task CreateOrderAsync(Order newOrder);
     }
