@@ -1,12 +1,11 @@
 ﻿using KittyStore.Domain.UserAggregate;
-using KittyStore.Domain.UserAggregate.ValueObjects;
 
 namespace KittyStore.Application.Common.Interfaces.Persistence
 {
     public interface IUserRepository
     {
         Task<List<User>> GetAllUsers();
-        Task<User?> GetUserByIdAsync(UserId id);
+        Task<User?> GetUserByIdAsync(Guid id);
         Task<User?> GetUserByEmailAsync(string email);
     
         Task AddUserAsync(User newUser);
