@@ -1,13 +1,11 @@
 ﻿using ErrorOr;
 using KittyStore.Domain.OrderAggregate;
-using KittyStore.Domain.UserAggregate.ValueObjects;
 using MediatR;
 
 namespace KittyStore.Application.Orders.Commands.CreateOrder
 {
     public record CreateOrderCommand (
-        AddressCommand AddressCommand,
-        UserId UserId) : IRequest<ErrorOr<Order>>;
+        AddressCommand AddressCommand) : IRequest<ErrorOr<Order>>;
 
     public record AddressCommand( 
         string Country, 
