@@ -1,12 +1,9 @@
 ﻿using ErrorOr;
-using KittyStore.Domain.CatAggregate.ValueObjects;
 using KittyStore.Domain.ShopCartAggregate;
-using KittyStore.Domain.UserAggregate.ValueObjects;
 using MediatR;
 
 namespace KittyStore.Application.ShopCarts.Commands.AddItem
 {
     public record AddShopCartItemCommand(
-        decimal Price,
-        CatId CatId) : IRequest<ErrorOr<ShopCart>>;
+        Guid CatId) : IRequest<ErrorOr<ShopCart>>;
 }
