@@ -1,10 +1,8 @@
 ﻿using ErrorOr;
 using KittyStore.Application.Common.Interfaces.Cache;
-using KittyStore.Application.Common.Interfaces.Persistence;
-using KittyStore.Application.ShopCarts.Common;
+using KittyStore.Application.Common.Interfaces.Utils;
 using KittyStore.Domain.Common.Errors;
 using KittyStore.Domain.ShopCartAggregate;
-using MapsterMapper;
 using MediatR;
 
 namespace KittyStore.Application.ShopCarts.Queries.GetShopCart
@@ -12,25 +10,26 @@ namespace KittyStore.Application.ShopCarts.Queries.GetShopCart
     public class GetShopCartQueryHandler : IRequestHandler<GetShopCartQuery, ErrorOr<ShopCart>>
     {
         private readonly ICacheService _cacheService;
-        private readonly IUserRepository _userRepository;
-        private readonly IMapper _mapper;
+        private readonly ICurrentUserService _currentUserService;
 
-        public GetShopCartQueryHandler(ICacheService cacheService, IUserRepository userRepository, IMapper mapper)
+        public GetShopCartQueryHandler(ICacheService cacheService,
+            ICurrentUserService currentUserService)
         {
             _cacheService = cacheService;
-            _userRepository = userRepository;
-            _mapper = mapper;
+            _currentUserService = currentUserService;
         }
 
         public async Task<ErrorOr<ShopCart>> Handle(GetShopCartQuery query, CancellationToken cancellationToken)
         {
-            if (await _userRepository.GetUserByIdAsync(query.UserId) is not { } user)
+            if (!_currentUserService.TryGetUserId(out var userId))
                 return Errors.User.NotFound;
-        
-            if(await _cacheService.GetDataAsync<ShopCartDto>(user.Id.ToString()) is not { } cart)
-                return Errors.ShopCart.ShopCartEmpty;
+
+            var cart = await _cacheService.GetDataAsync<ShopCart>(userId.ToString())
+                ?? ShopCart.Create(userId);
             
-            return _mapper.Map<ShopCart>(cart);
+            await _cacheService.SetDataAsync(userId.ToString(), cart,
+                DateTimeOffset.Now.AddDays(10));
+            return cart;
         }
     }
 }
