@@ -1,5 +1,6 @@
 ﻿using ErrorOr;
 using KittyStore.Application.Common.Interfaces.Persistence;
+using KittyStore.Application.Common.Interfaces.Utils;
 using KittyStore.Domain.Common.Errors;
 using KittyStore.Domain.UserAggregate;
 using MediatR;
@@ -9,15 +10,19 @@ namespace KittyStore.Application.Users.Commands.AddBalanceToUser
     public class AddBalanceToUserCommandHandler : IRequestHandler<AddBalanceToUserCommand, ErrorOr<User>>
     {
         private readonly IUserRepository _userRepository;
+        private readonly ICurrentUserService _currentUserService;
 
-        public AddBalanceToUserCommandHandler(IUserRepository userRepository)
+        public AddBalanceToUserCommandHandler(IUserRepository userRepository, 
+            ICurrentUserService currentUserService)
         {
             _userRepository = userRepository;
+            _currentUserService = currentUserService;
         }
 
         public async Task<ErrorOr<User>> Handle(AddBalanceToUserCommand command, CancellationToken cancellationToken)
         {
-            if (await _userRepository.GetUserByIdAsync(command.UserId) is not {} user)
+            var user = await _currentUserService.GetUserAsync();
+            if(user is null)
                 return Errors.User.NotFound;
 
             user.Balance.Replenishment(command.Amount);
