@@ -1,8 +1,7 @@
 ﻿using ErrorOr;
-using KittyStore.Domain.UserAggregate.ValueObjects;
 using MediatR;
 
 namespace KittyStore.Application.Users.Commands.DeleteUser
 {
-    public record DeleteUserCommand(UserId Id) : IRequest<ErrorOr<Unit>>;
+    public record DeleteUserCommand(Guid Id) : IRequest<ErrorOr<Unit>>;
 }
