@@ -1,13 +1,12 @@
 ﻿using ErrorOr;
 using KittyStore.Domain.UserAggregate;
-using KittyStore.Domain.UserAggregate.ValueObjects;
 using MediatR;
 
 namespace KittyStore.Application.Users.Commands.UpdateUser
 {
     public record UpdateUserCommand : IRequest<ErrorOr<User>>
     {
-        public UserId Id { get; set; } = default!;
+        public Guid Id { get; set; } = default!;
         public string FirstName { get; init; } = default!;
         public string LastName { get; init; } = default!;
         public string Email { get; init; } = default!;
