@@ -1,28 +1,27 @@
 ﻿namespace KittyStore.Domain.Common.Models
 {
-    public abstract class Entity<TId> : IEquatable<Entity<TId>>
-        where TId : notnull
+    public abstract class Entity : IEquatable<Entity>
     {
-        public TId Id { get;}
+        public Guid Id { get;}
 
-        protected Entity(TId id)
+        protected Entity(Guid id)
         {
             Id = id;
         }
     
-        public bool Equals(Entity<TId>? other) =>
+        public bool Equals(Entity? other) =>
             Equals((object?)other);
 
         public override bool Equals(object? obj) =>
-            obj is Entity<TId> entity && Id.Equals(entity.Id);
+            obj is Entity entity && Id.Equals(entity.Id);
 
         public override int GetHashCode() =>
             Id.GetHashCode();
 
-        public static bool operator ==(Entity<TId> left, Entity<TId> right) =>
+        public static bool operator ==(Entity left, Entity right) =>
             Equals(left, right);
     
-        public static bool operator !=(Entity<TId> left, Entity<TId>? right) =>
+        public static bool operator !=(Entity left, Entity? right) =>
             !Equals(left, right);
         
         #pragma warning disable CS8618
