@@ -1,17 +1,16 @@
 ﻿using KittyStore.Domain.Common.Models;
 using KittyStore.Domain.OrderAggregate.Entities;
 using KittyStore.Domain.OrderAggregate.ValueObjects;
-using KittyStore.Domain.UserAggregate.ValueObjects;
 
 namespace KittyStore.Domain.OrderAggregate
 {
-    public sealed class Order : AggregateRoot<OrderId>
+    public sealed class Order : AggregateRoot
     {
         private readonly List<OrderItem> _items = new();
     
         public Address Address { get; private set; }
     
-        public UserId UserId { get; private set;}
+        public Guid UserId { get; private set;}
     
         public DateTime Created { get; private set;}
 
@@ -19,7 +18,7 @@ namespace KittyStore.Domain.OrderAggregate
 
         public IReadOnlyList<OrderItem> OrderItems => _items;
 
-        private Order(OrderId id, Address address, UserId userId, decimal totalPrice, DateTime created) : base(id)
+        private Order(Guid id, Address address, Guid userId, decimal totalPrice, DateTime created) : base(id)
         {
             Address = address;
             UserId = userId;
@@ -27,10 +26,10 @@ namespace KittyStore.Domain.OrderAggregate
             TotalPrice = totalPrice;
         }
 
-        public static Order Create(Address address, decimal totalPrice, UserId userId) =>
-            new(OrderId.CreateUnique(), address, userId, totalPrice, DateTime.UtcNow);
+        public static Order Create(Address address, decimal totalPrice, Guid userId) =>
+            new(Guid.NewGuid(), address, userId, totalPrice, DateTime.UtcNow);
 
-        public void AddItems(List<OrderItem> items) => _items.AddRange(items);
+        public void AddItems(IEnumerable<OrderItem> items) => _items.AddRange(items);
         
         #pragma warning disable CS8618
             private Order() { }
