@@ -1,25 +1,25 @@
-﻿using KittyStore.Domain.CatAggregate.ValueObjects;
-using KittyStore.Domain.Common.Models;
-using KittyStore.Domain.ShopCartAggregate.ValueObjects;
+﻿using KittyStore.Domain.Common.Models;
+using Newtonsoft.Json;
 
 namespace KittyStore.Domain.ShopCartAggregate.Entities
 {
-    public sealed class ShopCartItem : Entity<ShopCartItemId>
+    public sealed class ShopCartItem : Entity
     {
         public decimal Price { get; private set;}
     
-        public CatId CatId { get; private set;}
+        public Guid CatId { get; private set;}
     
-        public ShopCartId ShopCartId { get; private set;}
-
-        public ShopCartItem(ShopCartItemId id, decimal price, CatId catId, ShopCartId shopCartId) : base(id)
+        public Guid ShopCartId { get; private set;}
+        
+        [JsonConstructor]
+        private ShopCartItem(Guid id, decimal price, Guid catId, Guid shopCartId) : base(id)
         {
             Price = price;
             CatId = catId;
             ShopCartId = shopCartId;
         }
 
-        public static ShopCartItem Create(decimal price, CatId catId, ShopCartId shopCartId) =>
-            new(ShopCartItemId.CreateUnique(), price, catId, shopCartId);
+        public static ShopCartItem Create(decimal price, Guid catId, Guid shopCartId) =>
+            new(Guid.NewGuid(), price, catId, shopCartId);
     }
 }
