@@ -17,9 +17,6 @@ namespace KittyStore.Infrastructure.Persistence.EntityConfigurations
         
             builder.Property(user => user.Id)
                 .ValueGeneratedNever()
-                .HasConversion(
-                    id => id.Value,
-                    value => UserId.Create(value))
                 .IsRequired();
 
             builder.Property(e => e.Balance)
@@ -41,7 +38,8 @@ namespace KittyStore.Infrastructure.Persistence.EntityConfigurations
             builder.Property(u => u.CreatedDateTime).IsRequired();
             builder.Property(u => u.UpdatedDateTime).IsRequired();
 
-            builder.HasData(SeedData.Users);
+            builder.HasData(
+               SeedData.CreateAdmin("Jorge", "Admin", "secret123", "admin123@gmail.com"));
         }
     }
 }
